+++ core/src/components/UnifiedSearch/SearchFilterChip.vue
