+++ core/src/services/UnifiedSearchService.js
@@ -1,10 +1,7 @@
 /**
- * @copyright 2020, John Molakvoæ <skjnldsv@protonmail.com>
+ * @copyright 2023, Fon E. Noel NFEBE <fenn25.fn@gmail.com>
  *
- * @author Christoph Wurst <christoph@winzerhof-wurst.at>
- * @author Daniel Calviño Sánchez <danxuliu@gmail.com>
- * @author Joas Schilling <coding@schilljs.com>
- * @author John Molakvoæ <skjnldsv@protonmail.com>
+ * @author Fon E. Noel NFEBE <fenn25.fn@gmail.com>
  *
  * @license AGPL-3.0-or-later
  *
@@ -23,16 +20,8 @@
  *
  */
 
-import { generateOcsUrl } from '@nextcloud/router'
-import { loadState } from '@nextcloud/initial-state'
+import { generateOcsUrl, generateUrl } from '@nextcloud/router'
 import axios from '@nextcloud/axios'
-
-export const defaultLimit = loadState('unified-search', 'limit-default')
-export const minSearchLength = loadState('unified-search', 'min-search-length', 1)
-export const enableLiveSearch = loadState('unified-search', 'live-search', true)
-
-export const regexFilterIn = /(^|\s)in:([a-z_-]+)/ig
-export const regexFilterNot = /(^|\s)-in:([a-z_-]+)/ig
 
 /**
  * Create a cancel token
@@ -46,7 +35,7 @@ const createCancelToken = () => axios.CancelToken.source()
  *
  * @return {Promise<Array>}
  */
-export async function getTypes() {
+export async function getProviders() {
 	try {
 		const { data } = await axios.get(generateOcsUrl('search/providers'), {
 			params: {
@@ -71,9 +60,13 @@ export async function getTypes() {
  * @param {string} options.type the type to search
  * @param {string} options.query the search
  * @param {number|string|undefined} options.cursor the offset for paginated searches
+ * @param {string} options.since the search
+ * @param {string} options.until the search
+ * @param {string} options.limit the search
+ * @param {string} options.person the search
  * @return {object} {request: Promise, cancel: Promise}
  */
-export function search({ type, query, cursor }) {
+export function search({ type, query, cursor, since, until, limit, person }) {
 	/**
 	 * Generate an axios cancel token
 	 */
@@ -84,6 +77,10 @@ export function search({ type, query, cursor }) {
 		params: {
 			term: query,
 			cursor,
+			since,
+			until,
+			limit,
+			person,
 			// Sending which location we're currently at
 			from: window.location.pathname.replace('/index.php', '') + window.location.search,
 		},
@@ -93,4 +90,18 @@ export function search({ type, query, cursor }) {
 		request,
 		cancel: cancelToken.cancel,
 	}
+}
+
+/**
+ * Get the list of active contacts
+ *
+ * @param {object} filter filter contacts by string
+ * @param filter.searchTerm
+ * @return {object} {request: Promise}
+ */
+export async function getContacts({ searchTerm }) {
+	const { data: { contacts } } = await axios.post(generateUrl('/contactsmenu/contacts'), {
+		filter: searchTerm,
+	})
+	return contacts
 }
