+++ core/templates/layout.user.php
@@ -68,7 +68,6 @@ p($theme->getTitle());
 			</div>
 
 			<div class="header-right">
-			    <div id="global-search"></div>
 				<div id="unified-search"></div>
 				<div id="notifications"></div>
 				<div id="contactsmenu"></div>
